refactor(core): move agent-state classification to agent-state.ts with an options object
New src/core/agent-state.ts owns AgentState type and classifyAgentState;
positional params replaced with {configured,proxyEnabled,installed,inRegistry}
options object; external-models.ts retains only dormancy/alias logic per ADR-013.
